+++ src/Configuration.jl
@@ -15,7 +15,7 @@ const TEST  = "test"
 haskey(ENV, "SEARCHLIGHT_ENV") || (ENV["SEARCHLIGHT_ENV"] = DEV)
 
 # defaults
-const SEARCHLIGHT_VERSION = v"0.12.1"
+const SEARCHLIGHT_VERSION = v"0.12.2"
 
 
 """