+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## v0.12.1 - 2019-08-29
 
+* bug fixes
+* deps update
+
+## v0.12.1 - 2019-08-29
+
 * SearchLight environment inherits Genie environment, if available
 * Improved handling of 2006 and 2013 MySQL errors (connection lost)
 * fixed an issue where reading the configuration file left the file open